create helper functions for tagger params for main agant and process agent
